requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
fpdf2==2.7.6
flask==3.0.0